+++ src/Brandlabs/Productiveio/Resources/Bookings.php
@@ -3,15 +3,25 @@
 namespace Brandlabs\Productiveio\Resources;
 
 use Brandlabs\Productiveio\ApiClient;
+use Brandlabs\Productiveio\BaseResource;
+use Brandlabs\Productiveio\Resources\Contracts\Crud;
+use Brandlabs\Productiveio\Resources\Contracts\GetList;
+use Brandlabs\Productiveio\Resources\Traits\ListResource;
+use Brandlabs\Productiveio\Resources\Traits\CreateResource;
+use Brandlabs\Productiveio\Resources\Traits\DeleteResource;
+use Brandlabs\Productiveio\Resources\Traits\GetResource;
+use Brandlabs\Productiveio\Resources\Traits\UpdateResource;
 
 /**
  * Productiveio Booking resource
  *
  * @package Brandlabs\Productiveio\Resources
  */
-class Booking extends BaseCrudResource
+class Bookings extends BaseResource implements Crud, GetList
 {
-    const RESOURCE_PATH = '/api/v2/bookings';
+    use CreateResource, DeleteResource, GetResource,ListResource, UpdateResource;
+
+    const RESOURCE_PATH = '/bookings';
 
     /**
      * Booking resource constructor