+++ src/Brandlabs/Productiveio/Resources/Companies.php
@@ -3,10 +3,21 @@
 namespace Brandlabs\Productiveio\Resources;
 
 use Brandlabs\Productiveio\ApiClient;
+use Brandlabs\Productiveio\BaseResource;
+use Brandlabs\Productiveio\Resources\Contracts\Create;
+use Brandlabs\Productiveio\Resources\Contracts\Get;
+use Brandlabs\Productiveio\Resources\Contracts\GetList;
+use Brandlabs\Productiveio\Resources\Contracts\Update;
+use Brandlabs\Productiveio\Resources\Traits\ListResource;
+use Brandlabs\Productiveio\Resources\Traits\CreateResource;
+use Brandlabs\Productiveio\Resources\Traits\GetResource;
+use Brandlabs\Productiveio\Resources\Traits\UpdateResource;
 
-class Companies extends BaseCrudResource
+class Companies extends BaseResource implements Create, Get, GetList, Update
 {
-    const RESOURCE_PATH = '/api/v2/companies';
+    use CreateResource, GetResource, ListResource, UpdateResource;
+
+    const RESOURCE_PATH = '/companies';
 
     public function __construct(ApiClient $apiClient)
     {