+++ src/Brandlabs/Productiveio/Resources/Contracts/Create.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Contracts;
+
+interface Create
+{
+		public function create(array $payload = []);
+}