+++ src/Brandlabs/Productiveio/Resources/Contracts/Crud.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Contracts;
+
+interface Crud extends Get, Create, Delete, Update
+{
+}