+++ src/Brandlabs/Productiveio/Resources/Contracts/Get.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Contracts;
+
+interface Get
+{
+	public function get(string $id);
+}