+++ src/Brandlabs/Productiveio/Resources/Contracts/GetList.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Contracts;
+
+interface GetList
+{
+	public function getList(
+        array $requestParams = [],
+        bool $aggregates = false
+    );
+}