+++ src/Brandlabs/Productiveio/Resources/Contracts/Update.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Contracts;
+
+interface Update
+{
+	public function update(string $id, array $payload = []);
+}