+++ src/Brandlabs/Productiveio/Resources/Tasks.php
@@ -3,10 +3,20 @@
 namespace Brandlabs\Productiveio\Resources;
 
 use Brandlabs\Productiveio\ApiClient;
+use Brandlabs\Productiveio\BaseResource;
+use Brandlabs\Productiveio\Resources\Contracts\Crud;
+use Brandlabs\Productiveio\Resources\Contracts\GetList;
+use Brandlabs\Productiveio\Resources\Traits\ListResource;
+use Brandlabs\Productiveio\Resources\Traits\CreateResource;
+use Brandlabs\Productiveio\Resources\Traits\DeleteResource;
+use Brandlabs\Productiveio\Resources\Traits\GetResource;
+use Brandlabs\Productiveio\Resources\Traits\UpdateResource;
 
-class Tasks extends BaseCrudResource
+class Tasks extends BaseResource implements Crud, GetList
 {
-    const RESOURCE_PATH = '/api/v2/tasks';
+    use CreateResource, DeleteResource, GetResource,ListResource, UpdateResource;
+
+    const RESOURCE_PATH = '/tasks';
 
     public function __construct(ApiClient $apiClient)
     {