+++ src/Brandlabs/Productiveio/Resources/Traits/GetResource.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Brandlabs\Productiveio\Resources\Traits;
+
+trait GetResource
+{
+    public function get(string $id)
+    {
+        return $this->getApiClient()->getResource($this->getEndpoint(), $id);
+    }
+}